Added the code required to create the "Help" tab next to "Navigate Terms" and "Find." (see #1314 for details)
The tab contains links (that open in a new tab) to useful pieces of
information and instructional videos, as well as details about how to
handle some of the less intuitive parts of the HERON interface. The
idea is to be able to provide users with immediate, visible answers to
common questions. A limited amount of CSS styling has been done, but
it could be improved. Currently the "Help" tab does not respond
properly when the "Show Options" or "Resize" buttons are clicked; the
tab should be more thoroughly tested to make sure it does not cause
any other bugs.

Most of the user-visible content is in default.htm under the tag with
ID "ontHelpDisp." Back-end code in ONT_view_Help.js is comprised of
code taken from ONT_view_Find.js and repurposed for the new
tab. Resolving the improper responses to the buttons mentioned above
will most likely involve more of the same procedure.

Dan C. created a makefile that allows for easy changes to the bmidev12
container. This streamlines the testing process by allowing the "make"
command to replace manually updating all edited files.

--HG--
branch : help_tab_1314